pyproject.toml
[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[project]
name = "boomgame"
version = "0.4.6"
description = "BOOM (MacOs shareware) remake with pygame"
readme = "README.md"
authors = [
    { name = "Raphael Reme", email = "raphaelreme-dev@protonmail.com" }
]
maintainers = [
    { name = "Raphael Reme", email = "raphaelreme-dev@protonmail.com" }
]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["BOOM", "PyGame", "Bomberman", "2D Game"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.9"
dependencies = [
    "importlib-resources ; python_full_version < '3.10'",
    "pygame-ce",
    "typing-extensions ; python_full_version < '3.11'",
]

[dependency-groups]
dev = [
    "ipywidgets",
    "matplotlib",
    "mypy",
    "notebook",
    "prek",
    "pygbag",
    "pyinstaller",
    "ruff",
]


[project.urls]
Homepage = "https://github.com/raphaelreme/boomgame"
Play = "https://raphaelreme.github.io/boomgame"
Documentation = "https://github.com/raphaelreme/boomgame"
Repository = "https://github.com/raphaelreme/boomgame"
Issues = "https://github.com/raphaelreme/boomgame/issues"

[project.scripts]
boom_version = "boomgame:display_version"

[project.gui-scripts]
BOOM = "boomgame.main:main"


[tool.uv]
python-preference = "only-managed"


[tool.ruff]
line-length = 120
indent-width = 4


# Format
format.docstring-code-format = true
format.indent-style = "space"
format.quote-style = "double"
format.line-ending = "lf"
format.skip-magic-trailing-comma = false

# Lint
lint.select = ["ALL"]

lint.ignore = [
    "ANN001",  # Checked by mypy which allows default args not to be typed
    "COM812",  # Recommended by ruff as we use the formatter
    "D100",  # Module docstring
    "D107",  # __init__ docstring
    "EM", "TRY003",  # Let's rely on standard exceptions
    "FBT",  # Ignore boolean trap
    "FIX", "TD",  # Ignore TODO, FIXME, XXX
    "S311",  # Pseudo-randomness is enough
    "S101",  # assert are ok for now. TODO: Remove assert.
]


lint.future-annotations = true
lint.flake8-annotations.ignore-fully-untyped = true
lint.flake8-annotations.mypy-init-return = true
lint.pep8-naming.ignore-names = [
    "i",
    "j",
    "k",
    "n",
    "t",
]
lint.pydocstyle.convention = "google"
lint.pylint.max-args = 8

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["INP001", "D103", "S101", "ANN201", "SLF001", "T201"]
"example/*.py" = ["INP001", "D103", "S101", "ANN201", "SLF001", "T201", "NPY002", "ERA001"]


[tool.mypy]
enable_error_code = ["ignore-without-code"]
check_untyped_defs = true
